requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.37
pydantic==2.10.5
python-dotenv==1.0.1
pika==1.3.2
python-multipart==0.0.20
psycopg2-binary==2.9.10
litellm==1.59.0
tenacity==9.0.0
pytz==2024.2